(xarcan) partial fixes to dri3/glamor
This fixes dirty-rectangles during rendering with -glamor, and with
upstream shmif patch - also fix so that the correct device will be
used on radeonsi based clients.

Accelerated clients still don't update correctly, and they don't
compose on the root node (likely related and somehow tied to what
we do with formats and so on in BO).