order translation adjusted
This is not entirely correct either, but will work for the current
unredirected case as the only thing that really matters is the
distinction between clipped and fully visibile toplevels. As long as the
fullyVisibile (unobscured) ones are assigned a higher order, the
translation done in the arcan composition side is fine.

When dynamic redirected backing stores is added, the ordering might need
to be more refined (use the clipping tree pattern again) but much easier
to test as alpha-blending is a better giveaway than the degenerate
surfaces that will have correct contents but incorrect ordering.