a different stab at tree synchronization
Not satisifed with the results from any of the previous attempts, time
to take a different take on things. It turns out that my understanding
of the arguments to ScreenStr:RestackWindow was flawed and the stack
that was forwarded was a shadow of what existed before. This explains a
lot of the ordering glitches we had before. This needs to be parried
with changes to the arcan/tests/xwm implementation, but hopefully things
will be smaller / cleaner.