add -wmexec, -miniwm and default to hwmouse
-wmexec [path] will have the [path] spawn with the DISPLAY env. setup,
intended to be used to autostart a WM inheriting from the Xarcan process.

The proper way to do this, alas, requires changes to xlib/xcb in order
for the display-open to not use the /tmp/.X0 style way of locating the
connection socket. When there's nothing left on the Xarcan todo, it
might be worth looking into that but for now there are other things to
fix.

The accelerated mouse cursor code seem to be mostly working, ARGB bits
are still untested and there are some cross-cutting issues with games
that explicitly grab+warp to implement a FPS style camera. Some tests
with Xonotic using changes to upstream arcan AND durden worked if we
force-set target/input/mouse/lock=center but the viewport based mouse
warp + new coordinates should 'in theory' work as well.

The big problem is the inherent race condition, before the warp takes
effect there might be unwarped samples in the pipe already, forcing
us to discard until it is applied.