wait for redirect segment req reply
Previous attempts at doing this deferred was not worth the pain,
switch to a different event processing loop to deal with the initial
segreq bursts.

There still is something with the redirected tree damageReport that
doesn't seem to hit right for how some WMs stack decorations, seem
that the last 'dirty' update synchs too early and then stalls until
invalidated by a configure.