switch redirect to work through pixmaps
The approach of making the shmif bindings on a per windows basis proved
to clash too much with COMPOSE and it turned out better to do it through
its pixmap indirection allocation. This comes with other downsides, but
better than to re-implement most of the extension.

The bigger caveat thus far is that there will be two shmif contexts in
use when a window is resized, one pre-resize and one post-resize.

These might not always synch on the arcan meta-wm side for a frame
and unless a better workaround is solved for, needs to flip between
the two vids on the 'resize' event for the new segment to guarantee
that the viewport-invisible event arrives at the same logical frame.

With synchronisation fences it might be possible to tie the viewport to
the fence it relates to and have the engine internal queue mask it all
to the WM. This still hurts network transparency in that there are
two in-flight channels momentarily and that they don't share compressor
state (though resizes typically enforce a flush of those regardless).

Other than that, there is a remaining ordering issue for the redirect
form, then need to fix -wmexec to mesh with the -miniwm for
single-client throwaway Xarcan instances. Lastly, some metadata synch
for the windows themselves and a clash between -redirect and -glamor and
we should be on par with arcan-wayland -exec-x11 but cleaner internally.