test: Xwayland doesn't start when another X server is running
When running the xserver:xwayland / XTS test in an environment where
an X server is running for :0, Xwayland fails to start with error:

    (EE)
    Fatal server error:
    (EE) Server is already active for display 0
        If this server is no longer running, remove /tmp/.X0-lock
        and start again.
    (EE)

And the `timeout` command fails instead of waiting for Weston to
initialize.

Add the `-displayfd` parameter to Xwayland to avoid this issue.

The `-displayfd` parameter lets Xwayland pick an unused display number,
avoiding the start up error when an X server is already active for :0,
the default display number.

Suggested-by: Michel Dänzer <mdaenzer@redhat.com>
Signed-off-by: José Expósito <jexposit@redhat.com>