(Xarcan) synch PRESENT, tune debug output, fakerel
This gets -redirect -glamor -present -exec glxgears "mostly" working but
with invalid contents. The grand problem is that wl-present as lifted is
made to work with the quirks of frame-callbacks which does not line up
with STEPFRAME events that well. What we would need is tracking between
flip, present and release, almost as if a locked opaque stream being
sampled would, yet again, be a cleaner interface than the fustercluck
years of yolo-patching managed.

While we can fake a depth of 3-4 and FIFO (or with synch handles,
MAILBOX) them, just releasing oldest on stepframe - what we really need
is explicit fencing with feedback on stepframe. The major problem with
that (outside the layer violations) is, as shown in recent ANDROID, that
FD passing with planes+buffers+synch has high overhead, and negotiating
a set which can then be cycled / signalled atomically as we did with
VREADY bitmap is the way to go - but good luck getting that into MESA
and the rest of the 'upside down pyramid' of client-device latched
allocations.