(dix/arcan) default-route OUTPUT segment
This is an incomplete form of a two step rocket. The first is to route
OUTPUT segments to a threaded proxy handler that connects back via xcb
and maps the shm context as xcb-shm.

The second stage to this is to mark a proxy-pixmap for externally
triggered copy requests for specific visuals (including the root
window).

This should help providing dynamic and controlled sharing with tools
that rely on copying window contents for sharing/recording.